+++ server/deckr/network/connection.py
@@ -5,16 +5,15 @@ This module provides code for deckr connections.
 import logging
 import traceback
 
-from twisted.protocols.basic import LineReceiver
-
-from google.protobuf.message import DecodeError
-from proto.client_message_pb2 import ClientMessage
-from proto.server_response_pb2 import ServerResponse
+import google.protobuf.message
+import proto.client_message_pb2
+import proto.server_response_pb2
+import twisted.protocols.basic
 
 LOGGER = logging.getLogger(__name__)
 
 
-class Connection(LineReceiver):
+class Connection(twisted.protocols.basic.LineReceiver):
     """
     This represents a single client connection. Each connection can be associated with at most
     1 game.
@@ -48,15 +47,15 @@ class Connection(LineReceiver):
                 and then pass off to the router.
         """
 
-        decoded_message = ClientMessage()
+        decoded_message = proto.client_message_pb2.ClientMessage()
         try:
             decoded_message.ParseFromString(message)
-        except DecodeError:
+        except google.protobuf.message.DecodeError:
             self.send_error("Could not parse message")
             return
 
         LOGGER.debug("Got a message %s from %s", decoded_message, self)
-        if decoded_message.message_type == ClientMessage.QUIT:
+        if decoded_message.message_type == proto.client_message_pb2.ClientMessage.QUIT:
             self.transport.loseConnection()  # twisted specific
             return
 
@@ -75,8 +74,8 @@ class Connection(LineReceiver):
         """
 
         LOGGER.warn("Sending error message %s", message)
-        response = ServerResponse()
-        response.response_type = ServerResponse.ERROR
+        response = proto.server_response_pb2.ServerResponse()
+        response.response_type = proto.server_response_pb2.ServerResponse.ERROR
         response.error_response.message = message
         self.send_response(response)
